requirements.txt
Django
Pillow
psycopg2-binary
python-decouple
drf_yasg
djangorestframework
celery
redis
requests
bs4
lxml
html5lib
asyncio
aiohttp
django-modeltranslation